% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADAGEpath.R
\docType{package}
\name{ADAGEpath}
\alias{ADAGEpath}
\alias{ADAGEpath-package}
\title{ADAGEpath: ADAGE-based Signature Analysis}
\description{
ADAGEpath provides functions to perform ADAGE-based signature
analysis on a gene expression dataset.
}
\details{
To learn more about ADAGEpath, read the README and start with the vignettes:
\code{browseVignettes(package = "ADAGEpath")}
}

