% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_network.R
\name{annotate_gene_signatures}
\alias{annotate_gene_signatures}
\title{Gene-signature relationship}
\usage{
annotate_gene_signatures(signatures_genes)
}
\arguments{
\item{signatures_genes}{a named list with each element storing genes in one
signature.}
}
\value{
a data.frame with the first column specifying geneID and the second
column specifying signatures that a gene is in.
}
\description{
Annotates genes with the signatures they are in.
}

