% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_network.R
\name{annotate_shared_signatures}
\alias{annotate_shared_signatures}
\title{Shared signature annotation}
\usage{
annotate_shared_signatures(gene_network_edge, signatures_genes)
}
\arguments{
\item{gene_network_edge}{a data.frame specifying gene-gene connections
in a gene-gene network. The first two columns are "to" and "from" geneIDs.
It is an intermediate product from the visualize_gene_network() function.}

\item{signatures_genes}{a named list with each element storing genes in one
signature.}
}
\value{
the data.frame gene_network_edge with one more column
shared_signatures that specifies the shared signatures of the from and to
genes.
}
\description{
Annotates shared signatures for gene pairs in the input gene_network_edge
data.frame.
}

