% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_signature.R
\name{annotate_signatures}
\alias{annotate_signatures}
\title{Signature annotation}
\usage{
annotate_signatures(signatures, annotation_terms, all_genes,
  significance_cutoff = 0.05)
}
\arguments{
\item{signatures}{a named list of signatures returned by the function
extract_signatures().}

\item{annotation_terms}{a named list of annotation terms returned by the
function fetch_geneset().}

\item{all_genes}{a vector storing all possible background genes in the
enrichment test.}

\item{significance_cutoff}{numeric, FDR significance cutoff used to filter
the test result (default to 0.05).}
}
\value{
a data.frame storing significantly enriched annotation terms for all
signatures.
}
\description{
Annotates signatures using GO or KEGG terms downloaded from Tribe. Annotation
is done through a gene set enrichment test (fisher exact test) with FDR
correction.
}

