% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_signature.R
\name{annotate_signatures_with_genesets}
\alias{annotate_signatures_with_genesets}
\title{Signatures annotation with gene sets}
\usage{
annotate_signatures_with_genesets(selected_signatures, model, genesets,
  significance_cutoff = 0.05)
}
\arguments{
\item{selected_signatures}{a character vector storing names of signatures}

\item{model}{an ADAGE model to extract signatures from}

\item{genesets}{a named list of gene sets returned by the
function fetch_geneset(), each element in the list is a character vector
storing genes (PAO1 locus tag) in a gene set.}

\item{significance_cutoff}{numeric, FDR significance cutoff used to filter
gene sets (default: 0.05).}
}
\value{
a data.frame storing significantly enriched gene sets for the
input signatures.
}
\description{
Annotate signatures with known GO or KEGG gene sets downloaded from the Tribe
webserver. Annotation is done through a gene set enrichment test (fisher
exact test) with FDR correcting the number of gene sets tested for each
signature.
}

