% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_signature.R
\name{build_gene_signature_map}
\alias{build_gene_signature_map}
\title{Gene-signature map}
\usage{
build_gene_signature_map(signatures_genes)
}
\arguments{
\item{signatures_genes}{a named list with each element storing genes in one
signature.}
}
\value{
a data.frame with the first column specifying geneID and the second
column specifying signatures that a gene is in.
}
\description{
Map genes to the signatures they are in.
}

