% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_signature.R
\name{build_signature_overlap_matrix}
\alias{build_signature_overlap_matrix}
\title{Signature overlap matrix construction}
\usage{
build_signature_overlap_matrix(signature_list, overlap_qvalues)
}
\arguments{
\item{signature_list}{a list of signatures extracted from an ADAGE model.}

\item{overlap_qvalues}{a list of adjusted p values obtained from the function
test_signature_overlap()}
}
\value{
a symmetric matrix storing the adjusted p values in the signature
overlap tests between every possible signature pair. The diagnol is set to
0 because a signature always has perfect overlap with itself.
}
\description{
Converts the list of adjusted p values in the signature overlap tests into
a symmetric matrix.
}

