% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_activity.R
\name{calculate_marginal_activity}
\alias{calculate_marginal_activity}
\title{Marginal activity calculation}
\usage{
calculate_marginal_activity(input_data, selected_signatures,
  model = eADAGEmodel)
}
\arguments{
\item{input_data}{a data.frame with gene IDs in the first column and
expression values from the second column.}

\item{selected_signatures}{a vector storing names of selected signatures}

\item{model}{an ADAGE model to extract signatures from
(default: the 300-node eADAGE model preloaded in the package).}
}
\value{
a data.frame storing marginal activities for the input samples.
It's rownames is set as "signature1-signature2",
indicating a marginal activity for signature1 after
removing the effect of signature2. If rowname is "signature1-signature1",
then it's the activity of signature1.
}
\description{
Calculates the marginal activities between all possible pairs of the input
signatures. Marginal activity is defined as the activity of a signature
after removing genes that overlap with another signature.
}

