% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ArrayExpress.R
\name{check_accession}
\alias{check_accession}
\title{Accession number validity check}
\usage{
check_accession(accession)
}
\arguments{
\item{accession}{ArrayExpress experiment accession number in the format of
"E-XXXX-n".}
}
\value{
logical, TRUE if the accession number has an associated experiment
and the experiment is on the "A-AFFY-30" platform; otherwise FALSE.
}
\description{
Checks whether the accession number has an associated experiment and the
experiment was measured on the Pseudomonas affymetric platform "A-AFFY-30".
}

