% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{check_input}
\alias{check_input}
\title{Input format check}
\usage{
check_input(input_data)
}
\arguments{
\item{input_data}{A data.frame.}
}
\value{
TRUE if the input_data.frame meets the requirements, otherwise FALSE.
}
\description{
Checks whether the input is a data.frame (or a tibble) with its first
column being character and the rest columns being numeric.
}

