% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_activity.R
\name{combine_geneset_outputs}
\alias{combine_geneset_outputs}
\title{Combining geneset outputs}
\usage{
combine_geneset_outputs(signature_geneset_association, geneset_limma_result)
}
\arguments{
\item{signature_geneset_association}{a data.frame storing signatures and
their significantly enriched gene sets, returned by the function
annotate_signatures_with_genesets().}

\item{geneset_limma_result}{a data.frame that stores the result table
returned by limma. It includes logFC, adj.P.Val, and other statistics for
each feature. The features are in the format "signature|geneset".}
}
\value{
a data.frame that stores both signature geneset association and
genesets' limma result.
}
\description{
Combines the signature geneset association output and the geneset limma
output.
}

