% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_network.R
\name{create_logFC_colors}
\alias{create_logFC_colors}
\title{Mapping logFC to HEX color}
\usage{
create_logFC_colors(logFC)
}
\arguments{
\item{logFC}{a numeric vector storing logFC values}
}
\value{
a character vector storing the HEX color values
}
\description{
Maps the continuous logFC values to HEX color values. The positive values
are mapped to the "Reds" palette from Color Brewer 2
\url{http://colorbrewer2.org/#type=sequential&scheme=Reds&n=3},
and the negative values
are oppsitely mapped to the "Blues" palette from the Color Brewer 2
\url{http://colorbrewer2.org/#type=sequential&scheme=Blues&n=3}.
}

