% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eADAGEmodel}
\alias{eADAGEmodel}
\title{eADAGE model}
\format{A data frame (tibble) with 5549 rows and 301 variables. The first
column is geneID that specifies gene identifiers. Starting from
the second column, each column stores numeric weight values of one node}
\source{
\url{https://doi.org/10.1101/078659}
}
\usage{
eADAGEmodel
}
\description{
The 300-node eADAGE model used in the paper "Unsupervised extraction of
functional gene expression signatures in the bacterial pathogen Pseudomonas
aeruginosa with eADAGE"
}
\keyword{datasets}

