% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_signature.R
\name{enrich_test}
\alias{enrich_test}
\title{Gene set enrichment test}
\usage{
enrich_test(set1, set2, set_all)
}
\arguments{
\item{set1}{character vector storing genes in the first gene set}

\item{set2}{character vector storing genes in the second gene set}

\item{set_all}{character vector storing all possible genes}
}
\value{
pvalue in the fisher exact test
}
\description{
Performs a one-side fisher exact test to determine whether two sets of genes
have significant overlap.
}

