% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{experimentID}
\alias{experimentID}
\title{Experiments and samples in Pseudomonas aeruginosa gene expression compendium}
\format{A data.frame with 1118 rows and 2 variables:
\describe{
 \item{Experiment}{ArrayExpress experiment assession numbers, only contain
 experiments in the preloaded PAcompendium data object.}
 \item{Sample}{Sample IDs in "xxx.CEL" format}
}}
\usage{
experimentID
}
\description{
ArrayExpress experiment assession numbers and their associated sample IDs
included in the preloaded PAcompendium data object.
There are duplicates in sample IDs because one sample can be included in
multiple experiments.
}
\keyword{datasets}

