% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expressionrange}
\alias{expressionrange}
\title{gene expression range}
\format{A data frame with 5549 rows and 3 variables:
\describe{
 \item{geneID}{P.a. PAO1 gene identifiers}
 \item{max_express}{Each gene's maximum expression value across the compendium}
 \item{min_express}{Each gene's minimum expression value across the compendium}
}}
\usage{
expressionrange
}
\description{
The expression range of each gene in the P. aeruginosa compendium.
}
\keyword{datasets}

