% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_signature.R
\name{extract_signatures}
\alias{extract_signatures}
\title{Signature extraction}
\usage{
extract_signatures(model = eADAGEmodel, HW_cutoff = 2.5,
  use_symbol = FALSE)
}
\arguments{
\item{model}{the ADAGE model to be used for extracting gene signatures
(default: the 300-node eADAGE model preloaded in the pacakge).}

\item{HW_cutoff}{number of standard deviations from mean in a node's weight
distribution to be considered as high-weight (default to 2.5). Only
high-weight genes are included in gene signatures.}

\item{use_symbol}{logical, whether the returned signatures use gene symbol
as gene identifiers.}
}
\value{
a named list with each element being a gene signature
}
\description{
Extracts all gene signatures from an ADAGE model.
}

