% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_signature.R
\name{find_associated_signatures}
\alias{find_associated_signatures}
\title{Associated signatures retrieval}
\usage{
find_associated_signatures(input_genesets, gene_set_list, model,
  significance_cutoff = 0.05, signature_limma_result = NULL)
}
\arguments{
\item{input_genesets}{a character vector storing gene set names, must be
found in the provided gene set list}

\item{gene_set_list}{a named list storing all possible gene sets with each
element being a vector of gene IDs.}

\item{model}{an ADAGE model to extract signatures from}

\item{significance_cutoff}{numeric, FDR significance cutoff used to filter
the result (default: 0.05).}

\item{signature_limma_result}{a data.frame that stores the result table
returned by limma. It includes logFC, adj.P.Val, and other statistics for
each signature.}
}
\value{
a list storing associated signatures for each input gene set
}
\description{
Returns the significantly associated signatures for the input genesets. If
signature_limma_result is provided, it will only return the significantly
associated signatures that achieves the lowest adjusted p-value in the
signature_limma_result.
}

