% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneinfo}
\alias{geneinfo}
\title{Gene information}
\format{A data.frame (tibble) with 5698 rows and 15 variables:
\describe{
  \item{\samp{#tax_id}}{a numeric vector, taxonomic id of P. aeruginosa}
  \item{\code{GeneID}}{a numeric vector, unique gene identifiers}
  \item{\code{Symbol}}{a character vector, gene symbols or gene names}
  \item{\code{LocusTag}}{a character vector, locus tags or PA numbers}
  \item{\code{Synonyms}}{a numeric vector}
  \item{\code{dbXrefs}}{a character vector}
 \item{\code{chromosome}}{a numeric vector}
  \item{\code{map_location}}{a numeric vector}
  \item{\code{description}}{a character vector}
  \item{\code{type_of_gene}}{a character vector}
  \item{\code{Symbol_from_nomenclature_authority}}{a numeric vector}
  \item{\code{Full_name_from_nomenclature_authority}}{a numeric vector}
  \item{\code{Nomenclature_status}}{a numeric vector}
  \item{\code{Other_designations}}{a numeric vector}
  \item{\code{Modification_date}}{a numeric vector}
}}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/gene/DATA/GENE_INFO/Archaea_Bacteria/Pseudomonas_aeruginosa_PAO1.gene_info.gz}
}
\usage{
geneinfo
}
\description{
Pseudomonas aeruginosa gene information curated by NCBI.
It was downloaded from NCBI ftp on Oct. 7 2016. It is used for
mapping PA numbers (LocusTag) to gene Symbol.
}
\keyword{datasets}

