% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_activation.R
\name{get_paretofront}
\alias{get_paretofront}
\title{Pareto fronts calculation}
\usage{
get_paretofront(test_result, N_fronts)
}
\arguments{
\item{test_result}{a data.frame included in the list returned from
the build_limma() function.}

\item{N_fronts}{int, number of pareto fronts}
}
\value{
a charactor storing all signatures made into the top N_fronts.
}
\description{
Pareto fronts calculation
}

