% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ArrayExpress.R
\name{get_sample_info}
\alias{get_sample_info}
\title{Querying sample information from ArrayExpress}
\usage{
get_sample_info(accession)
}
\arguments{
\item{accession}{ArrayExpress experiment accession number in the format of
"E-XXXX-n".}
}
\value{
a data.frame storing sample information in the SDRF file
}
\description{
Gets sample information of an experiment stored in the SDRF (Sample and Data
Relationship Format) file from ArrayExpress database.
}

