% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{impute_miss_values}
\alias{impute_miss_values}
\title{Missing values imputation}
\usage{
impute_miss_values(input_data, ref_data = PAcompendium)
}
\arguments{
\item{input_data}{a data.frame with gene IDs in the first column and
expression values from the second column.}

\item{ref_data}{a data.frame storing gene expression values to be used
as a reference dataset for calculating nearest neighbors. Should have more
samples than the input_data. The first columns of input_data and ref_data that
specify gene IDs should exactly be the same. (default: the P.a. gene
expression compendium).}
}
\value{
the input_data data.frame with missing values being filled
}
\description{
Imputes and fills missing values using k-nearest neighbor method. The 5
nearest neighbors of missing genes are calculated mainly using the reference
data and then are used to fill in missing values in the input_data. It uses
the impute.knn function from the impute bioconductor package.
}
\seealso{
\code{\link[impute]{impute.knn}}
}

