% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_network.R
\name{map_value_to_color}
\alias{map_value_to_color}
\title{Mapping value to HEX color}
\usage{
map_value_to_color(value)
}
\arguments{
\item{value}{a numeric vector storing values to map}
}
\value{
a character vector storing the mapped HEX color values
}
\description{
Maps continuous values to HEX color values. The positive values
are mapped to the "Reds" palette from Color Brewer 2
\url{http://colorbrewer2.org/#type=sequential&scheme=Reds&n=3},
and the negative values
are oppsitely mapped to the "Blues" palette from the Color Brewer 2
\url{http://colorbrewer2.org/#type=sequential&scheme=Blues&n=3}.
}

