% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_signature.R
\name{one_signature}
\alias{one_signature}
\title{One signature extraction}
\usage{
one_signature(node_weight, geneID, side, HW_cutoff = 2.5)
}
\arguments{
\item{node_weight}{a vector storing weight values of each gene to a node}

\item{geneID}{gene identifiers correspond to the weight vector}

\item{side}{character, "pos" or "neg"}

\item{HW_cutoff}{number of standard deviations from mean in a node's weight
distribution to be considered as high-weight (default to 2.5).}
}
\value{
a character vector storing genes in the signature defined by the
weight vector and the side.
}
\description{
Extracts a single gene signature from an ADAGE model.
}

