% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_activity.R
\name{one_signature_activity}
\alias{one_signature_activity}
\title{One signature activity}
\usage{
one_signature_activity(weight_matrix, express_matrix, node, side,
  HW_cutoff = 2.5)
}
\arguments{
\item{weight_matrix}{a data.matrix storing the weight matrix in an ADAGE model.}

\item{express_matrix}{a data.matrix storing gene expression values in a dataset}

\item{node}{a int ranging from 1 to number of columns in weight_matrix}

\item{side}{character, "pos" or "neg"}

\item{HW_cutoff}{number of standard deviations from mean in a node's weight
distribution to be considered as high-weight (default to 2.5).}
}
\value{
a vector storing activities of one signature across samples
}
\description{
Calculates activities for a specific signature in an ADAGE model.
}

