% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{operons}
\alias{operons}
\title{Pseudomonas aeruginosa PAO1 operons}
\format{A list with each element being a character vector that stores genes
in one operon. Locus tag (PAXXXX) is used as gene ID.}
\source{
\url{http://csbl.bmb.uga.edu/DOOR/index.php}
}
\usage{
operons
}
\description{
Pseudomonas aeruginosa PAO1 operon predictions obtained from the DOOR (
Database of prOkaryotic OpeRons) database on Dec. 12 2016.
}
\keyword{datasets}

