% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PAcompendium}
\alias{PAcompendium}
\title{Pseudomonas aeruginosa gene expression compendium}
\format{A data frame (tibble) with 5549 rows and 1052 variables. The first
column is geneID that specifies gene identifiers. Starting from
the second column, each column stores numeric expression values of one sample.}
\source{
\url{https://doi.org/10.1101/078659}
}
\usage{
PAcompendium
}
\description{
The compendium contains all P.a. microarray datasets measured on the
[Pae_G1a] Affymetric Pseudomonas aeruginosa array
platform from the ArrayExpress database as of 31 July 2015. The expression
values in the compendium have been background corrected and quantile
normalized.
}
\keyword{datasets}

