% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PAO1orthologs}
\alias{PAO1orthologs}
\title{Pseudomonas aeruginosa PAO1 strain orthologs}
\format{A data.frame (tibble) with 392139 rows and 8 variables:
\describe{
 \item{Strain (Query)}{Pseudomonas aeruginosa PAO1 (Reference)}
 \item{Locus Tag (Query)}{Locus tag for PAO1}
 \item{Description (Query)}{Description of the PAO1 gene}
 \item{Strain (Hit)}{Ortholog hit strain}
 \item{Locus Tag (Hit)}{Locus tag of the hit ortholog}
 \item{Description (Hit)}{Description of the hit ortholog}
 \item{Percent Identity}{Percentage of sequence identity}
 \item{Alignment Length}{Alignment length}
 ...
}}
\source{
\url{http://pseudomonas.com/downloads/pseudomonas/pgd_r_16_1/Pseudomonas_aeruginosa_PAO1_107/Pseudomonas_aeruginosa_PAO1_107_orthologs.txt}
}
\usage{
PAO1orthologs
}
\description{
Pseudomonas aeruginosa PAO1 strain ortholog predictions obtained from
http://pseudomonas.com on Oct. 10 2016.
It uses genes in PAO1 strain as query and finds
orthologs hit in other P.a. strains.
}
\keyword{datasets}

