% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_signature.R
\name{plot_signature_overlap}
\alias{plot_signature_overlap}
\title{Signature overlap plot}
\usage{
plot_signature_overlap(overlap_matrix, signatures = NULL)
}
\arguments{
\item{overlap_matrix}{a symmetric matrix storing adjusted p values between
each signature pair in the overlap test}

\item{signatures}{a character vector specifying which signatures to include
in the plot}
}
\description{
Plots the overlap significance between signature pairs using heatmap. Values
in the heatmap represent -log10(qvalue). A higher value indicates a more
significant overlap. The diagnol is set to the highest value in the selected
signature pairs for visualization purpose. The plot will be difficult to
see if including more than 50 signatures.
}

