% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_activity.R
\name{prepare_marginal_matrix}
\alias{prepare_marginal_matrix}
\title{Marginal significance matrix preparation}
\usage{
prepare_marginal_matrix(marginal_limma_result)
}
\arguments{
\item{marginal_limma_result}{a data.frame that stores the limma result table
returned by the build_limma() function when used on marginal activity.
It's rownames is in the format of "signature1-signature2".}
}
\value{
a matrix storing the adjusted p values from
the activation test when the effect of the column signature is removed from
the row signature. Values in the diagonal of the matrix are the activation
significance of signatures themselves.
}
\description{
Converts the limma result on margianl activity into a signatureN * signatureN
matrix format.
}

