% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probedistribution}
\alias{probedistribution}
\title{probe distribution}
\format{A numeric vector:}
\usage{
probedistribution
}
\description{
The quantile distribution of microarray probes used in building the
P. aeruginosa compendium. It will be used as a reference distribution to
quantile normalize new datasets.
}
\keyword{datasets}

