% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{process_celfiles}
\alias{process_celfiles}
\title{CEL files processing}
\usage{
process_celfiles(cel_folder, use_ref = TRUE,
  quantile_ref = probedistribution)
}
\arguments{
\item{cel_folder}{file path to the folder storing cel files.}

\item{use_ref}{a logical value indicating whether the probe normalization
should be done using a reference quantile distribution (default: TRUE).}

\item{quantile_ref}{a vector storing the reference quantile distribution
(default: the quantile distribution of probes used in
normalzing the P.a. gene expression compendium) Since the input microarray
data need to be normalzied to the processed compendium, the default
should not be changed in most cases.}
}
\value{
a data frame containing normalized gene expression values with geneID
in the first column and then each cel file in one column.
}
\description{
Processes microarray data in cel format. Only cel files measured on the
"Pae_G1a" platform will be processed. If using reference, the quantile
normalization of probes is based on a reference quantile
distribution.
}

