% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{quantile_norm}
\alias{quantile_norm}
\title{Quantile normalization}
\usage{
quantile_norm(input_data, use_ref = FALSE, ref_data = PAcompendium)
}
\arguments{
\item{input_data}{A data frame with gene IDs in the first column and
expression values from the second column.}

\item{use_ref}{A logical value indicating whether the normalization should
be done based on the reference data (default: FALSE).}

\item{ref_data}{A data frame storing gene expression values to be used
as a reference dataset for quantile normalization. The first columns of
input_data and ref_data that specify gene IDs should exactly be the same.
(default: the P.a. gene expression compendium).}
}
\value{
The input data frame after quantile normalization.
}
\description{
Performs quantile normalization on the input data. If using reference,
quantile normalization is done using the quantile distribution derived
from the reference data.
}
\seealso{
\code{\link[preprocessCore]{normalize.quantiles}},
\code{\link[preprocessCore]{normalize.quantiles.use.target}},
\code{\link[preprocessCore]{normalize.quantiles.determine.target}}
}

