% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_activity.R
\name{remove_redundant_signatures}
\alias{remove_redundant_signatures}
\title{Redundant signature removal}
\usage{
remove_redundant_signatures(marginal_limma_result, sig_cutoff = 0.05)
}
\arguments{
\item{marginal_limma_result}{a data.frame that stores the limma result table
returned by the build_limma() function when used on marginal activity.
It's rownames is in the format of "signature1-signature2".}

\item{sig_cutoff}{a numeric value used as the significance cutoff.
(default: 0.05)}
}
\value{
a vector storing remaining signatures after removing redundant oens.
}
\description{
Removes signatures whose activity changes are no longer significant after
removing genes overlapped with another signature.
}

