% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_activation.R
\name{set_phenotype}
\alias{set_phenotype}
\title{Defining sample phenotypes}
\usage{
set_phenotype(phenotypes)
}
\arguments{
\item{phenotypes}{a character with phenotypes separated by comma,
e.g. "wt,wt,wt,mt,mt,mt"}
}
\value{
a factor storing the input phenotypes
}
\description{
Defines the phenotype of each sample in a dataset.
}

