% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_activity.R
\name{signature_geneset_activity}
\alias{signature_geneset_activity}
\title{Calculating gene set activity inside signatures}
\usage{
signature_geneset_activity(signature_geneset_df, gene_set_list, model,
  input_data, HW_cutoff = 2.5)
}
\arguments{
\item{signature_geneset_df}{a data.frame with the first column specifying
signature names and second column specifying gene set names.}

\item{gene_set_list}{a named list storing all possible gene sets with each
element being a vector of gene IDs. It is used to extract genes IDs given
the name of a gene set.}

\item{model}{the ADAGE model to be used for calculating signature activity.}

\item{input_data}{a data.frame with gene IDs in the first column and
expression values from the second column.}

\item{HW_cutoff}{number of standard deviations away from mean in a node's
weight distribution to be considered as high-weight (default to 2.5).}
}
\value{
a data.frame with the first column being "signature|geneset" names and
the rest columns storing activities for every sample in the input_data.
}
\description{
Calculates activity of signatures only using genes in the provided gene sets
}

