% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{TDM_RNAseq}
\alias{TDM_RNAseq}
\title{RNAseq data normalization with TDM}
\usage{
TDM_RNAseq(input_data, ref_data = PAcompendium)
}
\arguments{
\item{input_data}{a data.frame with gene IDs in the first column and
expression values from the second column.}

\item{ref_data}{a data.frame storing gene expression values of a
compendium built from microarray data. It is used as the reference data in
TDM (default: the P.a. gene expression compendium).}
}
\value{
a data.frame storing TDM normalized gene expression values from the
input_data.
}
\description{
Performs Training Distribution Matching (TDM) on the input RNAseq data
to normalize RNAseq expression values to a comparable range of the reference
microarray data.
}
\seealso{
\url{https://github.com/greenelab/TDM}
}

