% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_signature.R
\name{test_signature_overlap}
\alias{test_signature_overlap}
\title{Signature overlap calculation}
\usage{
test_signature_overlap(signature_list)
}
\arguments{
\item{signature_list}{a list of signatures extracted from an ADAGE model.}
}
\value{
a named list storing adjusted p values for all signature overlap.
tests
}
\description{
Tests how significant all possible pairs of signatures in the input list
overlap with each other.
}

