% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{to_LocusTag}
\alias{to_LocusTag}
\title{Gene ID convertion}
\usage{
to_LocusTag(input_ID)
}
\arguments{
\item{input_ID}{the input gene ID (character).}
}
\value{
the corresponding PAO1 locus tag ("PAXXXX") for the input gene or NA
if the input gene is not recognized.
}
\description{
Converts the input gene ID to PAO1 locus tag that are used in ADAGE model.
It currently can convert gene IDs on AFFY chip, PAO1 gene symbols, and
PAO1 gene orthologs in other P.a. strains. It returns NA if the input is not
recognized as one of above.
}

