% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{to_symbol}
\alias{to_symbol}
\title{Converting locus tag to symbol}
\usage{
to_symbol(input_ID)
}
\arguments{
\item{input_ID}{character, a PAO1 gene locus tag, such as "PA0001".}
}
\value{
the converted gene symbol for the input locus tag.
}
\description{
Converts PAO1 gene locus tag to gene symbol. The input ID should be a
PAO1 locus tag. If it is not found in the PAO1 geneinfo database,
it will not be converted and the return value will be the same as the input.
}

