% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_network.R
\name{visualize_gene_network}
\alias{visualize_gene_network}
\title{Gene-gene network visualization}
\usage{
visualize_gene_network(selected_signatures, model = eADAGEmodel,
  cor_cutoff = 0.5, gene_logFC = NULL)
}
\arguments{
\item{selected_signatures}{a vector storing names of signatures whose genes
to include in the gene-gene network}

\item{model}{an ADAGE model to build gene-gene network from
(default: the 300-node eADAGE model preloaded in the pacakge).}

\item{cor_cutoff}{numeric, the correlation cutoff to decide whether an edge
between two genes exists in the network (default: 0.5).}

\item{gene_logFC}{a data.frame with two columns, one is geneID, the other is
logFC (optional, defautl:NULL). If not provided, the node color in the
gene-gene network is uniform.}
}
\description{
Builds and visualizes an ADAGE-derived gene-gene network of genes in the
selected signatures. In the network, each node represents a gene and two
nodes are linked by an edge if the correlation of the two connected genes in
the ADAGE model is higher than a certain cutoff. Edge width denotes the
correlation strength and edge color indicates positive (magenta) or negative
(green) correlation. If each gene's expression fold change is provided, it is
reflected by node color with red meaning high positive fold change and blue
meaning high negative fold change.
}

