% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_signature.R
\name{annotate_genes_in_signatures}
\alias{annotate_genes_in_signatures}
\title{Annotating genes in signatures}
\usage{
annotate_genes_in_signatures(selected_signatures, model = eADAGEmodel)
}
\arguments{
\item{selected_signatures}{a character vector storing names of signatures}

\item{model}{an ADAGE model to extract signatures from
(default: the 300-node eADAGE model preloaded in the package).}
}
\value{
a data.frame storing genes in the input signatures. Each gene is
annotated by gene symbol, gene description, and signatures it is in.
}
\description{
Annotates genes in the input signatures with their symbols, descriptions,
associated signatures.
}

