% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_activation.R
\name{build_limma}
\alias{build_limma}
\title{Linear model for activation detection}
\usage{
build_limma(activity, phenotypes, sample_indices = NULL)
}
\arguments{
\item{activity}{a data.frame that stores the signature activities for each
sample in a dataset.}

\item{phenotypes}{a factor obtained from the set_phenotype() function.}

\item{sample_indices}{a int vector to specify which samples to include
in the linear model.}
}
\value{
a list including test_result and phenotypes: test_result is a
data.frame that stores the activation differences and significance;
phenotypes is a factor after subseting the input phenotypes with
the sample_indices.
}
\description{
Uses limma to build a linear model of signature activities and test their
activations. Only two-group comparison is implemented for now.
}

