% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_signature.R
\name{build_signature_overlap_matrix}
\alias{build_signature_overlap_matrix}
\title{Signature overlap matrix construction}
\usage{
build_signature_overlap_matrix(selected_signatures, odds_ratios)
}
\arguments{
\item{selected_signatures}{a vector storing names of selected signatures.}

\item{odds_ratios}{a list of odds ratios obtained from the function
test_signature_overlap()}
}
\value{
a symmetric matrix storing the odds ratios in the signature
overlap tests between every possible signature pair. The diagnol is set to
Inf because a signature always has perfect overlap with itself.
}
\description{
Converts the list of odds ratios in the signature overlap tests into
a symmetric matrix.
}

