% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_activity.R
\name{calculate_activity}
\alias{calculate_activity}
\title{Signature activity calculation}
\usage{
calculate_activity(input_data, model = eADAGEmodel, HW_cutoff = 2.5)
}
\arguments{
\item{input_data}{a data frame with gene IDs in the first column and
expression values from the second column.}

\item{model}{the ADAGE model to be used for calculating signature activity
(default: the 300-node eADAGE model preloaded in the pacakge).}

\item{HW_cutoff}{number of standard deviations from mean in a node's weight
distribution to be considered as high-weight (default to 2.5).}
}
\value{
a named matrix storing activities per signature per sample
}
\description{
Calculates activities for each signature in an ADAGE model
specified by the weight matrix for each sample in the input data
}

