% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ArrayExpress.R
\name{download_raw}
\alias{download_raw}
\title{Downloading raw data from ArrayExpress}
\usage{
download_raw(accession, download_folder = "./")
}
\arguments{
\item{accession}{ArrayExpress experiment accession number in the format of
"E-XXXX-n".}
}
\value{
file path to the downloaded zip file that stores the cel files of
the input experiment.
}
\description{
Downloads the raw cel files of an experiment from the ArrayExpress database.
}

