% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eADAGEmodel}
\alias{eADAGEmodel}
\title{eADAGE model}
\format{A data.frame (tibble) with 5549 rows and 301 variables:
\describe{
  \item{geneID}{P.a. PAO1 gene identifiers}
  \item{Node1}{Each gene's weight to Node1}
  ...
}}
\source{
\url{http://biorxiv.org/content/early/2016/12/02/078659}
}
\usage{
eADAGEmodel
}
\description{
The 300-node eADAGE model used in the paper "Unsupervised extraction of
functional gene expression signatures in the bacterial pathogen Pseudomonas
aeruginosa with eADAGE"
}
\keyword{datasets}

