% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_signature.R
\name{fetch_geneset}
\alias{fetch_geneset}
\title{Gene set fetch}
\usage{
fetch_geneset(type = "KEGG", max_size = 100, min_size = 5)
}
\arguments{
\item{type}{character, type of the gene sets, either "GO" or "KEGG".}

\item{max_size}{int, maximum gene set size to be considered as a meaningful
gene set.}

\item{min_size}{int, minimum gene set size to be considered as a meaningful
gene set.}
}
\value{
a named list with each element being a gene set
}
\description{
Fetches GO or KEGG gene sets from the Tribe webserver
(url{tribe.greenelab.com}).
}

