% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_signature.R
\name{fetch_geneset}
\alias{fetch_geneset}
\title{Gene set fetch}
\usage{
fetch_geneset(type = "KEGG", username = NULL, access_date = NULL,
  max_size = Inf, min_size = 0)
}
\arguments{
\item{type}{character, type of the gene sets, can be "GO", "KEGG", or "User"
(default: "KEGG").}

\item{username}{character, creator's TRIBE username (default: NULL)}

\item{access_date}{character, in the format of month-day-year (12-31-16).
When specified, only the version of a term edited right before the access
date is retrieved. In default, a term's most up-to-date version is retrieved.
(default: NULL)}

\item{max_size}{int, maximum gene set size to be considered as a meaningful
gene set (default: Inf).}

\item{min_size}{int, minimum gene set size to be considered as a meaningful
gene set (default: 0).}
}
\value{
a named list with each element being a gene set
}
\description{
Fetches gene sets from the Tribe webserver
(url{https://tribe.greenelab.com}). It can retrieve GO, KEGG, or user-created
public gene sets for Pseudomonas aeruginosa.
}

