% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneinfo}
\alias{geneinfo}
\title{gene information}
\format{A data frame with 5698 rows and 15 variables:
\describe{
 \item{#taxid}{Taxonomic id of P. aeruginosa}
 \item{GeneID}{Unique gene identifiers}
 \item{Symbol}{Gene symbols or gene names}
 \item{LocusTag}{Locus tags or PA numbers}
 ...
}}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/gene/DATA/GENE_INFO/Archaea_Bacteria/Pseudomonas_aeruginosa_PAO1.gene_info.gz}
}
\usage{
geneinfo
}
\description{
Pseudomonas aeruginosa gene information curated by NCBI. It is used for
mapping PA numbers (LocusTag) to gene Symbol.
}
\keyword{datasets}

