% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_activation.R
\name{get_paretofront}
\alias{get_paretofront}
\title{Pareto fronts calculation}
\usage{
get_paretofront(input_data, N_fronts)
}
\arguments{
\item{input_data}{a data.frame with the first column storing a
solution's name and second and third columns storing values of its two
objectives.}

\item{N_fronts}{int, number of pareto fronts}
}
\value{
a charactor storing solutions made into the top N_fronts.
}
\description{
Returns solutions in the top N layers of pareto fronts with the goal to
maximize two objectives.
}

