% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_dataset}
\alias{load_dataset}
\title{Data loading and preprocessing}
\usage{
load_dataset(input, isProcessed = FALSE, isRNAseq = FALSE,
  model = eADAGEmodel, compendium = PAcompendium,
  quantile_ref = probedistribution, norm01 = TRUE)
}
\arguments{
\item{input}{file path to the input file or folder.}

\item{isProcessed}{a logical value indicating whether the input data has
already been processed into a tab-delimited txt file (default: FALSE).}

\item{isRNAseq}{a logical value indicating whether the processed input data
is RNAseq data. If true, the processed RNAseq data will be normalized to
a comparable range with the microarray-based compendium using TDM. If false,
the processed input data will be quantile normalized to be comparable
to the compendium. (default: FALSE)}

\item{model}{the ADAGE model used to analyze the input data
(default: the 300-node eADAGE model of P.a. pre-loaded within the package).}

\item{compendium}{the gene expression compendium of an organism
(default: the P. aeruginosa expression compendium pre-loaded within
the package).}

\item{quantile_ref}{a vector storing the reference quantile distribution
(default: the quantile distribution of probes used in
normalizing the P.a. gene expression compendium).}

\item{norm01}{a logical value indicating whether the output should be
zero-one normalized (default: TRUE)}
}
\value{
a data frame containing the processed gene expression values ready
for ADAGE analysis.
}
\description{
Currently supports three types of inputs:
a zip file containing microarray CEL files in a dataset (isProcessed=FALSE,
isRNAseq=FALSE);
a folder containing microarray CEL files in a dataset (isProcessed=FALSE,
isRNAseq=FALSE);
a tab-delimited txt file storing processed gene expression values in a
dataset (gene identifiers in the first column and then one sample per column)
(isProcessed=TRUE, isRNAseq=TRUE/FALSE).
}
\examples{
load_data(filepath, isProcessed = FALSE, isRNAseq = FALSE)
}

