% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_dataset}
\alias{load_dataset}
\title{Data loading and preprocessing}
\usage{
load_dataset(input, isProcessed = FALSE, isRNAseq = FALSE)
}
\arguments{
\item{input}{file path to the input file or folder.}

\item{isProcessed}{a logical value indicating whether the input data has
already been processed (default: FALSE).}

\item{isRNAseq}{a logical value indicating whether the input data is RNAseq
data. It determines whether TDM is applied to normalize the expression
values (default: FALSE).}
}
\value{
A data frame containing the processed gene expression values ready
for ADAGE analysis.
}
\description{
Currently supports three types of inputs:
a zip file containing microarray CEL files in a dataset;
a folder containing microarray CEL files in a dataset;
a tab-delimited txt file storing processed gene expression values in a
dataset (gene identifiers in the first column and then one sample per column).
}
\examples{
load_data(filepath, isProcessed = FALSE, isRNAseq = FALSE)
}

