% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_signature.R
\name{plot_signature_overlap}
\alias{plot_signature_overlap}
\title{Signature overlap plot}
\usage{
plot_signature_overlap(selected_signatures, model)
}
\arguments{
\item{selected_signatures}{a vector storing names of signatures to include
in the plot}

\item{model}{an ADAGE model to extract signatures from}
}
\description{
Plots the overlap significance between all pairs of signatures using a heatmap.
Values in the heatmap represent the odds ratio in fisher exact test on how
significant two signatures overlap in their genes.
A higher odds ratio indicates a more significant overlap between genes
in two signatures. The diagonal is set to the highest odds ratio in the
selected signature pairs just for visualization purpose. The plot will be
difficult to see if including more than 50 selected_signatures
}

