% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_signature.R
\name{plot_signature_overlap}
\alias{plot_signature_overlap}
\title{Signature overlap plot}
\usage{
plot_signature_overlap(selected_signatures, model = eADAGEmodel)
}
\arguments{
\item{selected_signatures}{a vector storing names of signatures to include
in the plot}

\item{model}{an ADAGE model to extract signatures from
(default: the 300-node eADAGE model preloaded in the pacakge).}
}
\description{
Plots the overlap significance between signature pairs using a heatmap.
Values in the heatmap represent the odds ratio in fisher exact test.
A higher odds ratio indicates a more significant overlap between genes
in two signatures. The diagnol is set to the highest odds ratio in the
selected signature pairs just for visualization purpose. The plot will be
difficult to see if including more than 50 signatures.
}

