% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{process_celfiles}
\alias{process_celfiles}
\title{Processing CEL files}
\usage{
process_celfiles(cel_folder, use_ref = TRUE,
  quantile_ref = probedistribution)
}
\arguments{
\item{cel_folder}{file path to the folder storing CEL files.}

\item{use_ref}{a logical value indicating whether the probe normalization
should be done using a reference quantile distribution (default: TRUE).}

\item{quantile_ref}{a vector storing the reference quantile distribution
(default: the quantile distribution of probes used in normalizing the P.a.
expression compendium pre-loaded within the package)}
}
\value{
a data.frame containing normalized gene expression values with geneID
in the first column and then each CEL file in one column.
}
\description{
Processes microarray data in CEL format. Only CEL files measured on the
"Pae_G1a" platform will be processed. If using reference, the quantile
normalization of probes is based on a reference quantile
distribution.
}

