% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{zeroone_norm}
\alias{zeroone_norm}
\title{Zero-one normalization}
\usage{
zeroone_norm(input_data, use_ref = FALSE, ref_data = compendium)
}
\arguments{
\item{input_data}{A data frame with gene IDs in the first column and
expression values from the second column.}

\item{use_ref}{A logical value indicating whether the normalization should
be done based on the reference data.}

\item{ref_data}{A data frame storing gene expression values to be used
as a reference dataset for zero-one normalization (default: the P.a. gene
expression compendium). The first columns in input_data and ref_data that
specify gene IDs should be exactly the same.}
}
\value{
A data frame storing zero-one normalized gene expression values from
the input data.
}
\description{
Normalize the input gene expression values to be between 0 and 1.
Normalization is done per gene (row) through
substracting its minimun from each value and then being divided by its range.
If using reference, the normalization is done using the gene's expression
mininum and range in the reference data.
}

